feat: implemented draw circle property
