+++ src/utils.ts
@@ -1,6 +1,7 @@
 type commandArray = [string, number]
 
 export const getParams = (params): commandArray => {
+	console.log(params.toString());
 	let [command, step] = params.toString().split(' ');
 
 	step = Number(step);
